gestor_de_medicamentos/core/serializers.py
from rest_framework import serializers

from .models import User, Medicamento


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = '__all__'


class MedicamentoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Medicamento
        fields = '__all__'
