requirements.txt
Django>=2.1.3,<3.0.0
djangorestframework>=3.9.0,<3.10.0
psycopg2>=2.7.5,<2.8.0
celery==5.1.0
redis==3.5.3

flake8>=3.6.0,<3.7.0
